Testy matematické knihovny a změna v matematické knihovně
Změna int na Byte v parametru funkce Fact.

Přidány testy funkcí matematické knihovny a úprava komentářů ve funkce Divide.

+++ src/CalculatorProject/MathLib/CalcMath.cs
@@ -48,7 +48,7 @@ namespace MathLib
     }
 
     // Vypočte faktorial z 'x'
-    public long Fact(int x)
+    public long Fact(Byte x)
     {
       return 0;
     }
